Extract "ddev init" functionalities into separate functions/files
