+++ src/commands/ddev/init.tsx
@@ -12,24 +12,23 @@ import { spawnInProcess } from "../../rendering/process/process_exec.js";
 import { Flags } from "@oclif/core";
 import { DDEVInitSuccess } from "../../rendering/react/components/DDEV/DDEVInitSuccess.js";
 import { DDEVConfig, ddevConfigToFlags } from "../../lib/ddev/config.js";
-import { hasBinary } from "../../lib/hasbin.js";
 import { ProcessRenderer } from "../../rendering/process/process.js";
 import { renderDDEVConfig } from "../../lib/ddev/config_render.js";
 import { loadDDEVConfig } from "../../lib/ddev/config_loader.js";
 import { Value } from "../../rendering/react/components/Value.js";
 import { ddevFlags } from "../../lib/ddev/flags.js";
-import { exec } from "child_process";
-import { promisify } from "util";
 import { compareSemVer } from "semver-parser";
 import { assertStatus, type MittwaldAPIV2 } from "@mittwald/api-client";
-import { Text } from "ink";
 import { readApiToken } from "../../lib/auth/token.js";
 import { isNotFound } from "../../lib/fsutil.js";
 import { dump, load } from "js-yaml";
+import { determineDDEVDatabaseId } from "../../lib/ddev/init_database.js";
+import {
+  assertDDEVIsInstalled,
+  determineDDEVVersion,
+} from "../../lib/ddev/init_assert.js";
 
 type AppInstallation = MittwaldAPIV2.Components.Schemas.AppAppInstallation;
-
-const execAsync = promisify(exec);
 
 export class Init extends ExecRenderBaseCommand<typeof Init, void> {
   static summary = "Initialize a new ddev project in the current directory.";
@@ -73,9 +72,14 @@ export class Init extends ExecRenderBaseCommand<typeof Init, void> {
 
     await assertDDEVIsInstalled(r);
 
-    const ddevVersion = await this.determineDDEVVersion(r);
+    const ddevVersion = await determineDDEVVersion(r);
     const appInstallation = await this.getAppInstallation(r, appInstallationId);
-    const databaseId = await this.determineDatabaseId(r, appInstallation);
+    const databaseId = await determineDDEVDatabaseId(
+      r,
+      this.apiClient,
+      this.flags,
+      appInstallation,
+    );
     await this.writeAuthConfiguration(r);
     const config = await this.writeMittwaldConfiguration(
       r,
@@ -100,15 +104,6 @@ export class Init extends ExecRenderBaseCommand<typeof Init, void> {
       "auth",
       "ssh",
     ]);
-  }
-
-  private async determineDDEVVersion(r: ProcessRenderer): Promise<string> {
-    const { stdout } = await execAsync("ddev --version");
-    const version = stdout.trim().replace(/^ddev version +/, "");
-
-    r.addInfo(<InfoDDEVVersion version={version} />);
-
-    return version;
   }
 
   private async getAppInstallation(
@@ -168,62 +163,6 @@ export class Init extends ExecRenderBaseCommand<typeof Init, void> {
     }
 
     return await r.addInput("Enter the project name", false);
-  }
-
-  private async determineDatabaseId(
-    r: ProcessRenderer,
-    appInstallation: AppInstallation,
-  ): Promise<string | undefined> {
-    let databaseId: string | undefined = this.flags["database-id"];
-    const withoutDatabase = this.flags["without-database"];
-
-    if (withoutDatabase) {
-      return undefined;
-    }
-
-    if (databaseId === undefined) {
-      databaseId = (appInstallation.linkedDatabases ?? []).find(
-        (db) => db.purpose === "primary",
-      )?.databaseId;
-    }
-
-    if (databaseId !== undefined) {
-      const mysqlDatabaseResponse =
-        await this.apiClient.database.getMysqlDatabase({
-          mysqlDatabaseId: databaseId,
-        });
-      assertStatus(mysqlDatabaseResponse, 200);
-
-      r.addInfo(<InfoDatabase name={mysqlDatabaseResponse.data.name} />);
-      return mysqlDatabaseResponse.data.name;
-    }
-
-    return await this.promptDatabaseFromUser(r, appInstallation);
-  }
-
-  private async promptDatabaseFromUser(
-    r: ProcessRenderer,
-    appInstallation: AppInstallation,
-  ): Promise<string | undefined> {
-    const { projectId } = appInstallation;
-    if (!projectId) {
-      throw new Error("app installation has no project ID");
-    }
-
-    const mysqlDatabaseResponse =
-      await this.apiClient.database.listMysqlDatabases({ projectId });
-    assertStatus(mysqlDatabaseResponse, 200);
-
-    return await r.addSelect("select the database to use", [
-      ...mysqlDatabaseResponse.data.map((db) => ({
-        value: db.name,
-        label: `${db.name} (${db.description})`,
-      })),
-      {
-        value: undefined,
-        label: "no database",
-      },
-    ]);
   }
 
   /**
@@ -298,14 +237,6 @@ export class Init extends ExecRenderBaseCommand<typeof Init, void> {
   }
 }
 
-async function assertDDEVIsInstalled(r: ProcessRenderer): Promise<void> {
-  await r.runStep("check if DDEV is installed", async () => {
-    if (!(await hasBinary("ddev"))) {
-      throw new Error("this command requires DDEV to be installed");
-    }
-  });
-}
-
 async function writeContentsToFile(
   filename: string,
   data: string,
@@ -321,21 +252,5 @@ function InfoUsingExistingName({ name }: { name: string }) {
     <>
       using existing project name: <Value>{name}</Value>
     </>
-  );
-}
-
-function InfoDDEVVersion({ version }: { version: string }) {
-  return (
-    <>
-      detected DDEV version: <Value>{version}</Value>
-    </>
-  );
-}
-
-function InfoDatabase({ name }: { name: string }) {
-  return (
-    <Text>
-      using database: <Value>{name}</Value>
-    </Text>
   );
 }
