+++ src/lib/auth/token.ts
@@ -3,10 +3,20 @@ import fs from "fs/promises";
 import { Config } from "@oclif/core";
 import { isNotFound } from "../fsutil.js";
 
+/**
+ * Gets the filename in which the CLI should store the API token.
+ *
+ * @param config The CLI configuration object.
+ */
 export function getTokenFilename(config: Config): string {
   return path.join(config.configDir, "token");
 }
 
+/**
+ * Reads the API token from the environment or the configuration file.
+ *
+ * @param config The CLI configuration object.
+ */
 export async function readApiToken(
   config: Config,
 ): Promise<string | undefined> {
@@ -15,6 +25,7 @@ export async function readApiToken(
   );
 }
 
+/** Reads the API token from the MITTWALD_API_TOKEN environment variable. */
 function readApiTokenFromEnvironment(): string | undefined {
   const token = process.env.MITTWALD_API_TOKEN;
   if (token === undefined) {
@@ -23,6 +34,11 @@ function readApiTokenFromEnvironment(): string | undefined {
   return token.trim();
 }
 
+/**
+ * Reads the API token from the configuration file.
+ *
+ * @param config The CLI configuration object.
+ */
 async function readApiTokenFromConfig(
   config: Config,
 ): Promise<string | undefined> {
